+++ vitest.config.ts
@@ -10,6 +10,6 @@ export default defineConfig({
   test: {
     clearMocks: true,
     environment: "node",
-    include: ["tests/**/*.test.ts"],
+    include: ["tests/**/*.test.{ts,mjs}"],
   },
 });